EPM-787
- removed redundant instructions from Unit 8 Assessment
- I think I clarified the Unit 10 Assessment instructions but could use
a second set of eyes. It seems like the goal of this exercise was to
have the students create the divs using js instead of html, and have
this happening in a for loop within a function. Happy to edit if that
wasn’t the intent or if you see any errors.